requirements.txt
flask
requests
pillow
torch
torchvision